Adds skin copying and listbox deletions, Updates UI to include JMs
Skin maps are now copied to a subdirectory where they are renamed to preserve the original maps. The user can now delete items in the maps and JMs listboxes using the '-' button under the respective LB. The UI is now updated to use the 'place()' method so more elements could be added.